refactor: creating a prop for name on FormRadio
