+++ src/formInput/index.ts
@@ -1,2 +1,2 @@
-export { FormInput, ErrorPosition } from './FormInput';
+export { FormInput } from './FormInput';
 export { FormInputMasked } from './FormInputMasked';
