+++ stories/5.FormRadioExplanation.stories.mdx
@@ -31,7 +31,6 @@ An example with all the available properties is:
   id='radio-id'
   inputProps={{
     className: 'class-name',
-    name: 'radio-name',
     onChange: handleChange
   }}
   itemProps={{
@@ -40,6 +39,7 @@ An example with all the available properties is:
   labelProps={{
     style: {display: 'flex'}
   }}
+  name='radio-name'
   selected={true}
 />
 ```
