+++ stories/6.FormRadio.stories.mdx
@@ -11,11 +11,9 @@ import "./style.css";
       return (
         <FormRadio
           label="text example"
+          name="group1"
           value="value"
           id="radio-1"
-          inputProps={{
-            name: "group1"
-          }}
         />
       );
     }}
