+++ stories/7.FormRadioStyled.stories.mdx
@@ -26,12 +26,12 @@ In this section we're using the formRadio component providing the GovUk style pa
             }}
             inputProps={{
               className: "govuk-radios__input",
-              name: "changed-name",
               onChange: handleChange
             }}
             itemProps={{
               className: "govuk-radios__item"
             }}
+            name="changed-name"
           />
           <FormRadio
             id="changed-name"
@@ -42,12 +42,12 @@ In this section we're using the formRadio component providing the GovUk style pa
             }}
             inputProps={{
               className: "govuk-radios__input",
-              name: "changed-name",
               onChange: handleChange
             }}
             itemProps={{
               className: "govuk-radios__item"
             }}
+            name="changed-name"
           />
         </div>
       )
@@ -72,12 +72,12 @@ In this section we're using the formRadio component providing the GovUk style pa
             }}
             inputProps={{
               className: "govuk-radios__input",
-              name: "changed-country",
               onChange: handleChange
             }}
             itemProps={{
               className: "govuk-radios__item"
             }}
+            name="changed-country"
           />
           <FormRadio
             id="changed-country"
@@ -88,12 +88,12 @@ In this section we're using the formRadio component providing the GovUk style pa
             }}
             inputProps={{
               className: "govuk-radios__input",
-              name: "changed-country",
               onChange: handleChange
             }}
             itemProps={{
               className: "govuk-radios__item"
             }}
+            name="changed-country"
           />
           <FormRadio
             id="changed-country"
@@ -104,12 +104,12 @@ In this section we're using the formRadio component providing the GovUk style pa
             }}
             inputProps={{
               className: "govuk-radios__input",
-              name: "changed-country",
               onChange: handleChange
             }}
             itemProps={{
               className: "govuk-radios__item"
             }}
+            name="changed-country"
           />
           <FormRadio
             id="changed-country"
@@ -120,12 +120,12 @@ In this section we're using the formRadio component providing the GovUk style pa
             }}
             inputProps={{
               className: "govuk-radios__input",
-              name: "changed-country",
               onChange: handleChange
             }}
             itemProps={{
               className: "govuk-radios__item"
             }}
+            name="changed-country"
           />
         </div>
       )
@@ -150,12 +150,12 @@ In this section we're using the formRadio component providing the GovUk style pa
             }}
             inputProps={{
               className: "govuk-radios__input",
-              name: "changed-name-disabled",
               onChange: handleChange
             }}
             itemProps={{
               className: "govuk-radios__item"
             }}
+            name="changed-name-disabled"
           />
           <FormRadio
             id="changed-name-disabled"
@@ -166,12 +166,12 @@ In this section we're using the formRadio component providing the GovUk style pa
             }}
             inputProps={{
               className: "govuk-radios__input",
-              name: "changed-name-disabled",
               onChange: handleChange
             }}
             itemProps={{
               className: "govuk-radios__item"
             }}
+            name="changed-name-disabled"
           />
           <FormRadio
             id="changed-name-disabled"
@@ -182,13 +182,13 @@ In this section we're using the formRadio component providing the GovUk style pa
             }}
             inputProps={{
               className: "govuk-radios__input",
-              name: "changed-name-disabled",
               onChange: handleChange
             }}
             itemProps={{
               className: "govuk-radios__item"
             }}
             disabled={true}
+            name="changed-name-disabled"
           />
         </div>
       )
@@ -213,13 +213,13 @@ In this section we're using the formRadio component providing the GovUk style pa
             }}
             inputProps={{
               className: "govuk-radios__input",
-              name: "changed-name-preselected",
               onChange: handleChange
             }}
             itemProps={{
               className: "govuk-radios__item"
             }}
             selected={true}
+            name="changed-name-preselected"
           />
           <FormRadio
             id="changed-name-preselected"
@@ -230,12 +230,12 @@ In this section we're using the formRadio component providing the GovUk style pa
             }}
             inputProps={{
               className: "govuk-radios__input",
-              name: "changed-name-preselected",
               onChange: handleChange
             }}
             itemProps={{
               className: "govuk-radios__item"
             }}
+            name="changed-name-preselected"
           />
         </div>
       )
@@ -260,7 +260,6 @@ In this section we're using the formRadio component providing the GovUk style pa
             }}
             inputProps={{
               className: "govuk-radios__input",
-              name: "changed-name-with-hint",
               onChange: handleChange
             }}
             itemProps={{
@@ -271,6 +270,7 @@ In this section we're using the formRadio component providing the GovUk style pa
               classes: "govuk-hint govuk-radios__hint",
               id: "sign-in-item-hint"
             }}
+            name="changed-name-with-hint"
           />
           <FormRadio
             id="changed-name-with-hint-2"
@@ -281,7 +281,6 @@ In this section we're using the formRadio component providing the GovUk style pa
             }}
             inputProps={{
               className: "govuk-radios__input",
-              name: "changed-name-with-hint",
               onChange: handleChange
             }}
             itemProps={{
@@ -292,6 +291,7 @@ In this section we're using the formRadio component providing the GovUk style pa
               classes: "govuk-hint govuk-radios__hint",
               id: "sign-in-2-item-hint"
             }}
+            name="changed-name-with-hint"
           />
         </div>
       )
